+++ src/types.ts
@@ -129,6 +129,8 @@ export declare type CaseFileOption = "k" | CaseOption;
 
 export interface AutoOptions {
   additional?: any;
+  additionalTable?: any;
+  additionalRelation?: string[];
   /** Case of file names */
   caseFile?: CaseFileOption;
   /** Case of model names */