+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # sequelize-auto change log
 
+## [0.8.9] - 2025-06-30
+
+* 2 new options bellow only support `Programmatic API`
+* Add `additionalRelation` option to support put self-defined relations to init-Model file
+* Add `additionalTable` option to support define per-table options. This options will overwrite `additional` option
+
 ## [0.8.7] - 2021-12-08
 
 * Don't write field comments for mssql, because it throws error in sequelize.sync() (#567)